[Refactor] Not Found in core
Moved file as it was common.
References #59 and #62.
