+++ server/services/delivery-service/src/interfaces/controllers/index.ts
@@ -1,7 +1,7 @@
 import {connectDrone, disconnectDrone} from '../../usecases';
 import makePostDrone from './post-drone';
 import makeDeleteDrone from './delete-drone';
-import notFound from './not-found';
+import notFound from '../../../../core/interfaces/controllers/not-found';
 
 const postDrone = makePostDrone({connectDrone});
 const deleteDrone = makeDeleteDrone({disconnectDrone});
