+++ server/services/order-service/src/interfaces/controllers/index.ts
@@ -1,7 +1,7 @@
 import {retrieveOrder, createOrder} from '../../usecases';
 import makeGetOrder from './get-order';
 import makePostOrder from './post-order';
-import notFound from './not-found';
+import notFound from '../../../../core/interfaces/controllers/not-found';
 
 const getOrder = makeGetOrder({retrieveOrder});
 const postOrder = makePostOrder({createOrder});
